Rename state-reader into state-api
- rename to conform with Python and as a setup to addding `trait State`.
- fixed typo in `get_class_hash_at` variable, doesn't need a leading
  underscore.

commit-id:461c1762
