+++ blockifier/src/state.rs
@@ -1,3 +1,3 @@
 pub mod cached_state;
 pub mod errors;
-pub mod state_reader;
+pub mod state_api;
